Avatars and Countries use the directory listing from the relevant graphics folders to fill the lists.
This makes adding and removing avatars and countries a simple matter of adding a .gif to the avatars folder or a .bmp file to the countries.

Existing database users with the country "No country" do not have a relevant flag now as it's been renamed to No Country for consistency with the menus.