Begin the work of moving functions and included files into an include folder.
Included files should give no output.  That's not 100% true at this time.  But the work is beginning.  Most of the includes will be functions that can be called by the main application to return the desired output or result.  This should lower the number of files in the main folder and make it easier to understand where all of the files fit into the application structure.

countires.inc.php has not been turned into a function at this point in time.  I wanted to commit the work done.  Other files have been altered to support the new pathnames and access methods to include files where appropriate.