Big cleanup of the Admin interface.  It's all too confusing to have so many subdirectories with files to manage when using the same header.  Better named files in the one Admin folder are easier to manage.  News was move in this fashion.  The entire interface was skinned to give the same look as the main site.  It reuses the same include files from the main site.
The elo functions were moved from multiple files into a single class in the main include folder.  Not all elo calculating files are using the include at the moment.  This commit is primarily for the Admin interface.

Login was revamped to use sessions properly.  With a security check to ensure you are logged in so you can use the Admin interface.

General code tidy and cleanup was performed on most files.  I got sick of the long time it was taking to make all the required changes to get this commit to be one large atomic commit, so I haven't completed all the cleanup that is required in the files.  This will need to be revisited at some point.

The Admin interface does not perform certain functions and they were removed.  You cannot alter the settings any longer.  They are all in the variables.conf file.  Also profile.php appeared to be a copy of the main profile.php and was therefore deleted. elo_func.inc.php was also never referenced.

The settings table is not used in the database, neither is the pages table.  Changes to the install were made to not create these redundant tables.