Move the user blocking funcationality into the admin section.  Most other functions are duplicates of what the user can do themselves.  This is better implemented by allowing admins to masquerade as users.  There is only one code path to maintain that way.
Maintenance is completed automatically and it not an admin function, it's been removed.  Not all of the changes that affect the use of maintenance are committed in this commit.  So initiall it will look like a loss of funcationality.

The menu is reworked to have groups at the top of the admin section, News and Admins, rather than individual functions.  All news functions can be done from within the news section now.

An admin listing has been added to the Admin Users screen.

Currently the Admin area has been changed to authenticate from the players tables.  There production database does not support this at this time, and will need to have it added to release this version.  The SQL to complete this part of the transformation is;

--
ALTER TABLE webl_players ADD COLUMN is_admin boolean DEFAULT false;
update webl_players SET is_admin = true WHERE name IN (select name from webl_admin)
DROP TABLE webl_admin;
--