----- Config file addition:
define("NONPUBLIC_REPLAY_COMMENTS", 0); //set to 1 to enable, 0 to 
disable. If enabled only logged in members can view other members game 
comments.


-----  Code:

Winners can now comment & sportsrate when reporting and then only, on 
purpose to encourage doing it directly and since they alredy are 
fiddling with the page. Loser can do it whenever from the profile, but a 
time limit will be implemneted in the future. Once a comment/rating has 
been set it can't be changed, also on purpose and to avoid frakked up 
conduct.