Added cookie abilities to jquery.
Alter profile to use ajax and allow expansion and contraction of recent games and available to play.  The code was cleaned a little, $row was replaced by a more meaningful $player.  As always, this page remembers your selections, so if you expand recent games it will stay expanded when you review a different profile.