Feedback have been moved to a game detail page.  All comments, the ability to contest, withdraw are all available on that page.
The gametable information is used in a few places and was placed in a function in the include folder.  It makes displaying the same table in multiple locations a lot easier.  Profile has been cleaned up to use game details as well.

The faq has had an invalid tag altered.

undogame was removed as it's all rolled into the game details now.