-----  Fixes:
- error in ladder.php isn't displayed any more when db is empty
- error in sidebar.php isn't displayed any more when db is empty

----- Added to config:

// Group ladder config
define("GADDER_ELO_RANGE", 50); // The range in points of a rating that 
a player needs to be within the group leader in order to be rated as 
within the same group on the gadder. It works the following way: The 
1:st group is based on the number one ranked player on the ladder. All 
who are within GADDER_ELO_RANGE of his rating are ranked as belonging to 
his group (the first one). As soon as a player is < GADDER_ELO_RANGE he 
will be placed in the next group. Summarised: Let the value reflect the 
point system and K-values in use and experiment with it to see if the 
Gadder is meaningfull. If it's not and groups too many together, then 
either your ladder is new with few games/players or you have set a too 
huge value here. Else, if the Gadder has very many groups, you have 
probably set the value too low: Keep in mind it's supposed to group 
players which it thinks are nearby or equal skillwise, saying that 
they're pretty much the same rank. E.g. a good default value when using 
standard Elo system with usual K and entry of 1500 would be somewhere 
between 45 - 65. This value however depends only on all the other 
settings made and is only valid with such parameters. Also notice that 
the gadder works the very same way as the ladder, with the obvious 
exception for the grouping and displaying of rank. Whoever is displayed 
in the normal ladder would appear somewhere in the gadder.
define("GADDER_FLAGS", 0); // Should we show flags in the gadder? 1 for 
yes, 0 for no. Personally I find it chaotic having it enabled.

----- Added:
- gadder.php, should be complete and do what it's supposed to.