1. Finally fixed so retarded spammers can't report thousand games within 5 minutes. Now a user has to wait x minutes between game reports, and once a user has reported y amount of games this feature will be disabled.
2. Also made it so the annoying verification pop-up after reporting a 
game is only displayed for players that have repored < x amount of 
games.
 

Added the following to config file:

define("SPAM_REPORT_TIME_PROTECTION", 20); // Number of minutes a user 
has to wait inbetween reporting a victory. 
define("SPAM_REPORT_TIME_PROTECTION_UNLOCKED", 100); // When a user has 
_reported_ at least these many games he won't adhere under the spam 
report protection rule. Instead he'll be able to report results within 
any time intervall. Set this number very high to shut off this feature. 
Set it to 50 or 100 or so to give the veterans the freedom.