Update: The number 400 in the Elo formula that regulates win expectancy can now be changed with ease from within the config file. Add this to the old config file:
define("ELO_DIVIDE_FACTOR",400); // This number regulates the win 
expectancy and is a part of the Elo formula. Leave it on 400 in full 
information games with 0 randomness like chess. Make it 800 or so in 
games with some more randomness in them. The more randomness, the higher 
the number should be, but, the higher the number and the more 
randomness, the more useless the Elo system gets.